brownfield-ai/requirements.txt
fastapi
uvicorn
langchain
openai
tiktoken
python-dotenv
httpx
pydantic
tenacity
gitpython
langchain-community
